Improve email address unobfuscation (Issue #17)
This converts all current email addresses on the MH list to the correct
version of themselves. However, it is dependent on what people choose
to write in the email address field, so it could break if someone else
decides to get creative.
I also added an email address validation before all the mail server
stuff, to avoid doing any work if the address is empty anyway.